requirements.txt
requests
python-dotenv
pytest
black
isort
flake8